requirements.txt
Flask==2.1.2
numpy==1.22.3
Pillow==9.0.0
tensorflow==2.8.0
